Our Iris project
Iris is a software made for tracking the movements of of eye the main tracked thing in eye is pupil and blink count. using pupil movements and blink count we r able to do a particular task in pc for a instance we add a task in it of if you blink your eye for two times it will increase 20% of brightness in your device. well you are able to add desire type of task in it if u know how to code  product made by me and my team by scratch from scratch.....